Validation des contraintes Inscription et desistement OK + tri tableau principal trié par date de sortie

+++ src/Controller/MainController.php
@@ -2,8 +2,6 @@
 
 namespace App\Controller;
 
-
-
 use App\Entity\Inscription;
 use App\Entity\Sortie;
 use App\Entity\SortiesSearch;
@@ -45,32 +43,34 @@ class MainController extends AbstractController
      */
     public function inscription( $id, EntityManagerInterface $em, Request $request )
     {
+
         $inscription = new Inscription();
+        $this->date = new \DateTime();
 
         $sortie = $this->getDoctrine()->getRepository(Sortie::class)->find($id);
+        $participants= $sortie->getParticipants();
+
         $participant = $this->getUser();
 
-        //$date = new \DateTime();
-        //dump($date);
+        //Contrainte => Un User ne peut pas s'inscrire 2 fois à la même sortie
+        foreach ( $participants as $ins) {
+            if($ins->getParticipant()->getId() == $this->getUser()->getId()){
+                $this->addFlash("danger", " ⚠ Impossible de vous inscrire 2 fois à la même sortie ! ⚠" );
+                return $this->redirectToRoute('outing_show',['id'=>$sortie->getId()]);
+            }
+        }
 
-        //Contrainte => Ne peut pas s'inscrire 2 fois à la même sortie
-        /*if($this->date > $sortie->getDateLimiteInscription()){
-            $this->addFlash("danger", "Impossible de vous inscrire car la dâte de clôture des inscriptions est dépassée !" );
+        //Contrainte => Ni forcer une inscription si le nombre de places maximales est atteint
+       if(count($sortie->getParticipants()) >= $sortie->getNbInscriptionsMax()){
+            $this->addFlash("danger", " ⚠ Impossible de vous inscrire car il n'y plus de places disponibles ! ⚠");
             return $this->redirectToRoute('outing_show',['id'=>$sortie->getId()]);
-        }*/
+        }
 
-
-        //Contrainte => Date de clôture inscription dépassée
-        /*if($this->date > $sortie->getDateLimiteInscription()){
-            $this->addFlash("danger", "Impossible de vous inscrire car la dâte de clôture des inscriptions est dépassée !" );
+        //Contrainte => Ni forcer une inscription si la Date de clôture inscription dépassée
+        if($this->date > $sortie->getDateLimiteInscription()){
+            $this->addFlash("danger", " ⚠ Impossible de vous inscrire car la dâte de clôture des inscriptions est dépassée ! ⚠" );
             return $this->redirectToRoute('outing_show',['id'=>$sortie->getId()]);
-        }*/
-
-        //Contrainte => NB de places maximales atteintes
-       /*if($sortie->getParticipants() >= $sortie->getNbInscriptionsMax()){
-            $this->addFlash("danger", "Impossible de vous inscrire car il n'y plus de places disponibles ! ");
-            return $this->redirectToRoute('outing_show',['id'=>$sortie->getId()]);
-        }*/
+        }
 
         $inscription->setParticipant($participant);
         $inscription->setSortie($sortie);
@@ -91,12 +91,17 @@ class MainController extends AbstractController
     {
         $inscription = $this->getDoctrine()->getRepository(Inscription::class)->findOneBy(['id' => $id]);
 
-        /*if($inscription->getParticipant() != $this->getUser()){
-            $this->addFlash("success", "Impossible de se désister, vous n'êtes pas encore inscrit !");
-            return $this->redirectToRoute('outing_show',[
-                'id'=>$inscription->getSortie()->getId()
-            ]);
-        }*/
+        //si l'objet Inscription est null
+        if( $inscription == null){
+            $this->addFlash("danger", " ⚠ Cette inscription n'existe pas, le désistement a échoué ! ⚠");
+            return $this->redirectToRoute('home');
+        }
+
+        //si le user tente un forcage d'un desistement par l'url à une sortie à laquelle il n'est pas inscrit
+        if( $inscription->getParticipant() != $this->getUser()){
+            $this->addFlash("danger", " ⚠ Impossible de se désister, vous n'êtes pas encore inscrit ! ⚠");
+            return $this->redirectToRoute('home');
+        }
 
         $em->remove($inscription);
         $em->flush();
@@ -106,6 +111,4 @@ class MainController extends AbstractController
             'id'=>$inscription->getSortie()->getId()
         ]);
     }
-
-
 }
